requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==2.0.30
aiomysql
python-dotenv
cryptography

pytest
pytest-asyncio
httpx
python-dotenv
